Minor fixes and test data added.
All that is needed for proper subscription demonstration is to load PSW needed .sql scripts, place breakpoints @ subscription controller and service impl., and SubscriptionScheduler.cs + BloodSubscriberConsumer.cs in psw.
After that just run the docker, isa, psw and wait for the triggers.
Ps. copy of test data can be found on discord.